requirements.txt
asgiref==3.5.2
autoflake==2.0.2
black==23.3.0
click==8.1.3
Django==4.1.1
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
isort==5.12.0
Markdown==3.4.1
mypy==1.2.0
mypy-extensions==1.0.0
mysql-client==0.0.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pyflakes==3.0.1
pytz==2022.5
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.5.0
zipp==3.10.0
